Re-factor stack_tags*.go to remove duplicate code.
 - stack_tags.go contains common code, useable in both Go and GopherJS runtime environments
 - stack_tags_js.go contains the GopherJS/Node.js specific runtime stack handling
 - stack_tags_main.go contains the standard Go runtime specific stack handling
